Added zeroconf plugin. The plugin uses zeroconf to find peers on the local netwok.
The zeroconf code was developed by Lesly Weyts and Kevin Andre as a school project. Had to rework it a bit because they based their implementation upon 2.0.3, and there have been some changes in the area of PeerSources since 2.0.3. 

Note: The plugin uses avahi, it is probably better to move to KDE's dnssd implementation in the future. But that is something for the 2.2 release of ktorrent or maybe even 3.0 (port to KDE4)


svn path=/trunk/extragear/network/ktorrent/; revision=616952