Increase maximum number of slots for magnet queue
The magnet queue was implemented to alleviate the network congestion
produced when a big list of magnets are downloaded at the same
time. Also, having so many magnets downloading at the same time was
pointed (prematurely and incorrectly, my fault) as the reason for some
hangs that KTorrent has sometimes when it are downloading such big list
of magnets.

As the number of concurrent downloading magnets are not directly related
with the hangs, this patch increases the number of downloading slots for
the queue manager from 20 to 100. So now, anyone can overload his
network if he want ;).