Improve performance of several Feed methods
A list of loaded items was being used to track the status of the Feed
items. Lists has O(n) complexity, leading to bad performance in the
methods that were used mainly to check if a item was contained in it.

This patch replaces the list for a set, leading to O(1) complexity. It
also adds a local cache of feed items ids stored in a set, that has the
same performance improvement purpose while checking the validity of the
already loaded items.