Port away from queryExit(), which is about to be deprecated.
KTorrent was using queryExit to:
1) save the state of the mainwindow for the next start
2) shut down the core

and I can see it had some problems with queryExit being called more than
once, although I'm not sure how/why this happened.

This patch takes care of 1) when closing the window and when quitting the app
(which can happen without closing the window, i.e. we remember that it was
visible), and of 2) by connecting to QCoreApp::aboutToQuit directly (which makes
"Core" more standalone in that respect).