Remove cross scripting
Cross is not well maintained, even before the Qt5 port.
For KF5 it is marked as porting aid. Considering that there
are no known users of this scripting and there is already a DBus
API it was agreed on removing it.

If there comes up the need for more scripting options one could extend the DBus
API or implement sth. in QJSEngine.

Task: https://phabricator.kde.org/T11608