Port to new KPluginFactory constructor
Also make the base class constructor compatible with KPluginFactory.
This is what is done in the KDE frameworks that provide plugin interfaces too.

Futhermore it sets the plugin metadata for the kpart. This will be useful in
case the plugin wants to access it's own metadata or if KPart will use it inernally.