Disable logging by default and add --enable-logging command line option.
Logging is not useful during normal operation of the program and only
generates unnecessary writes to the filesystem and periodic CPU load
when the log files are compressed. So disable logging by default.

Provide the --enable-logging command line option to enable logging
back. The option is now also required to enable logging on console
(which is further enabled by --verbose).

CCBUG: 430355