Fix behaviour when closing KTorrent to system tray
KStatusNotifierIcon hides the QWindow, while GUI hides the main QWidget.
It seems like hiding both of these was causing issues where the QWidget
was considered visible but the QWindow was considered hidden (i.e. clicking
on the tray icon did not close the window when it was visible)

Test-plan:
- Enable KTorrent's system tray icon
- Close the main KTorrent window
- Click the system tray icon twice

Previously:
- First click shows the window
- Second click does nothing

Now:
- First click shows the window
- Second click hides the window