Fix group view not being loaded after ktorrent is exited when closed to tray
When closing KTorrent to the tray we were hiding the window and then saving
the window state. This meant GroupView was always considered as not visible,
so the group view was not open by default when launching KTorrent again.

Now we save the window state before hiding, and only if the window is actually
visible.

Also there is no need to save the state in GUI::quit since QCoreApplication::quit
now attempts to close in windows in Qt 6, so queryClose is called anyway.