Save magnet queue whenever the queue is updated
I.e. when magnets are added to the queue. Previously the queue was only
saved when KTorrent was closed, so all magnets added in a session were lost
if KTorrent crashed.

Test-plan: import a stale magnet link then send a SIGHUP to KTorrent and
restart the app. Previously the magnet link did not appear on the
magnets list after the restart, now it does.

BUG: 415580