Undo system tray bug workaround
In KF5 KStatusNotifierIcon would show/hide the main QWindow whilst in KTorrent
we would show/hide the main QWidget. A bug in Qt6 meant that hiding/showing
the QWindow did not neccessarily hide the QWidget.

So if a user closed the window to the tray, the QWidget was considered hidden
but the QWindow was not. Then when clicking on the tray icon, the QWindow
visiblity was being toggled but the QWidget visibility was always hidden.

To work around this we changed KTorrent to hide the QWindow instead of the
QWidget. Now KStatusNotifierIcon has been updated to hide the main QWidget
instead of the QWindow, so we must undo the changes made here.

The upstream bug should fixed in Qt 6.7 anyway.

Reverts e35b4881605f045ebe8bca2da357c2a3ae2f1e02

Reverts efe4aa0dcc23e57e68ecd809068f1a49d7705a07

BUG: 484598