Fix crash when video widget closes
show_video_action::toggled is connected to showVideo, which calls
closeVideo. So when we set show_video_action::checked in closeVideo,
this caused closeVideo to be called recursively. Then we set video to be
a nullptr and return back to the first calling of closeVideo. Then we
try to dereference video again, which results in a crash.

The video widget still crashes for me anyway, but now due to the
phonon-vlc backend.

BUG: 490616