core: deprecate loadSilently in favour of load with similar args
`kt::TorrentInterface` function `loadSilently` is mostly copied code as the `load` function with same arguments.

All usages of this `loadSilently` function are also next to the corresponding `load` function with no other difference than the condition to load silently.

Refactor these 2 functions into 1, for less redundancy and 1 reduced "if" statement in favour of an extra argument.